+++ recipes/configure-elk-nginx-proxy.rb
@@ -4,7 +4,7 @@
 ::Chef::Recipe.send(:include, MhOpsworksRecipes::RecipeHelpers)
 
 elk_info = get_elk_info
-http_auth = elk_info[:http_auth]
+http_auth = elk_info['http_auth']
 es_host = node[:opsworks][:instance][:private_ip]
 
 template %Q|/etc/nginx/sites-enabled/default| do
@@ -20,7 +20,7 @@ end
 
 bash "htpasswd" do
   code <<-EOH
-    htpasswd -bc /etc/nginx/conf.d/kibana.htpasswd #{http_auth[:user]} #{http_auth[:pass]}
+    htpasswd -bc /etc/nginx/conf.d/kibana.htpasswd #{http_auth['user']} #{http_auth['pass']}
   EOH
 end